abstract out buffer geometries
