+++ ui/src/utils/three.ts
@@ -1,4 +1,8 @@
-import type { GeometricTriangle, GeometricParallelogram } from './render/geometric';
+import type {
+  GeometricTriangle,
+  GeometricParallelogram,
+  GeometricBilinearPatch,
+} from './render/geometric';
 
 export type TriangleGeometry = {
   vertices: Float32Array;
@@ -120,4 +124,86 @@ export function createParallelogramGeometry(
   const uvs = new Float32Array([0, 0, 1, 0, 0, 1, 1, 1]);
 
   return { vertices, indices, normals, uvs };
+}
+
+export type BilinearPatchGeometry = {
+  vertices: Float32Array;
+  indices: Uint16Array;
+  normals: Float32Array;
+  uvs: Float32Array;
+};
+
+/**
+ * returns raw vertex/index/normal data for a bilinear patch,
+ * subdivided into a 16x16 grid with normals from partial derivatives.
+ */
+export function createBilinearPatchGeometry(
+  geometricData: GeometricBilinearPatch,
+): BilinearPatchGeometry {
+  const { p00, p10, p01, p11 } = geometricData;
+
+  const divs = 16;
+  const positions: number[] = [];
+  const normals: number[] = [];
+  const indices: number[] = [];
+  const uvs: number[] = [];
+
+  const evalPatch = (u: number, v: number): [number, number, number] => {
+    const w00 = (1 - u) * (1 - v);
+    const w10 = u * (1 - v);
+    const w01 = (1 - u) * v;
+    const w11 = u * v;
+    return [
+      w00 * p00[0] + w10 * p10[0] + w01 * p01[0] + w11 * p11[0],
+      w00 * p00[1] + w10 * p10[1] + w01 * p01[1] + w11 * p11[1],
+      w00 * p00[2] + w10 * p10[2] + w01 * p01[2] + w11 * p11[2],
+    ];
+  };
+
+  // generate vertices and normals (divs+1 x divs+1 grid)
+  for (let j = 0; j <= divs; j++) {
+    const v = j / divs;
+    for (let i = 0; i <= divs; i++) {
+      const u = i / divs;
+      const p = evalPatch(u, v);
+      positions.push(...p);
+
+      // compute normal from partial derivatives
+      const eps = 0.001;
+      const du_u = Math.min(u + eps, 1.0 + eps);
+      const dv_v = Math.min(v + eps, 1.0 + eps);
+      const du = evalPatch(du_u, v);
+      const dv = evalPatch(u, dv_v);
+
+      const dp_du = [du[0] - p[0], du[1] - p[1], du[2] - p[2]];
+      const dp_dv = [dv[0] - p[0], dv[1] - p[1], dv[2] - p[2]];
+
+      // cross product dp_du x dp_dv
+      const nx = dp_du[1] * dp_dv[2] - dp_du[2] * dp_dv[1];
+      const ny = dp_du[2] * dp_dv[0] - dp_du[0] * dp_dv[2];
+      const nz = dp_du[0] * dp_dv[1] - dp_du[1] * dp_dv[0];
+      const len = Math.sqrt(nx * nx + ny * ny + nz * nz) || 1;
+      normals.push(nx / len, ny / len, nz / len);
+      uvs.push(u, v);
+    }
+  }
+
+  // generate triangle indices (two per grid cell)
+  for (let j = 0; j < divs; j++) {
+    for (let i = 0; i < divs; i++) {
+      const a = j * (divs + 1) + i;
+      const b = a + 1;
+      const c = a + (divs + 1);
+      const d = c + 1;
+      indices.push(a, b, d);
+      indices.push(a, d, c);
+    }
+  }
+
+  return {
+    vertices: new Float32Array(positions),
+    indices: new Uint16Array(indices),
+    normals: new Float32Array(normals),
+    uvs: new Float32Array(uvs),
+  };
 }
